Optimize menu file prop handling
Instantiating a menu file prop now reads all the data into a memory
stream backed by a shared byte array. This seems to perform a lot faster
than using the AFileBaseStream. When applying material texture changes,
the original texture that gets replaced is destroyed. When destroying
the prop controller, the materials, textures, and mesh are destroyed
manually.

OpenStream is not thread-safe since the cached backing array's data
would be overwritten and so MenuFileParser still uses the
AFileBaseStream.