Refactor IKDragHandleController.IKEnabled
IKEnabled had way too much control on a drag handle being enabled or
not. If I wanted some drag handles to be enabled even if IK is disabled,
it would not let me because Enabled had been overridden to only be true
if IK was also enabled.

Instead, have IKEnabled just be a flag that drag handle controllers can
use to decide how they want to handle that state.

This change also brings back being able to select a character's
body/face even if IK is disabled.