Fix animation time changing on character switched
The animation pane update's the animation position slider's right bound
using the Right property which will update the value with a notification
when switching characters. When the slider bound is changed from a
higher value to one that is lower than the slider's current value, the
slider's value will be clamped to the right bound and update the current
character's animation position. This causes unintended changes to
current character's pose as if the slider were being manipulated by the
user.