Improve method of finding preset on refresh
When a custom preset repository is refreshed, the selector panes will
try to update the dropdowns to match the character's current preset. If
the current preset's directory still exists, the selector will attempt
to find the preset within that category by comparing the ID. If the
preset cannot be found then it'll just set it to the first item in the
list.

When the category doesn't exist anymore the preset pane will default to
the first category and the first item but now without changing the
character's current preset.

This commit improves the check by checking for ID instead of relying on
the overrided equals method because the Equals method considers two
presets equal if the category and name match as well as the ID. This is
not ideal because the current preset could have been renamed.