Clean up BasePane/BaseWindow
Pane and window have changed a lot throughout development and had things
added and changed as complexity increased. Many of the things that
BaseWindow introduced have been pulled up into BasePane.

A clear separation has also been made between a window and a pane. A
BasePane could have other BasePanes added to it but the same is not
expected of BaseWindows. Some BaseWindows that exist now draw panes in
sequence, some have panes assigned in key value pairs and choose to draw
their panes in however way it needs. Making BaseWindow not implement
IEnumerable<BasePane> in order to add support for collection
initializers removes ambiguity on how the window can be initialized.