Refactor main window
Main window had this separate pane called a TabsPane that would be
passed to the BasePanes of the main window and those panes would draw
the tabs pane themselves. I don't know why it was done this way to begin
with but that nonsense has been removed. The main window took some
lessons from the new settings window and now draws the tabs itself.

Constants.Window has also been slimmed down to remove the main window
tabs as they are not windows at all and have instead been moved into the
MainWindow as MainWindow.Tab.