Disable UI when modal window is visible
BasePane had an Enabled property for the longest time but has never been
used. The property is now moved into BaseWindow and used by BasePanes to
enable drawing only if the BaseWindow is enabled and a modal window
isn't open.

This needed to be implemented because I'm not using GUI.Modal anymore
which handled other windows interactivity while open. See 438b6b5