Fix dictionary keys serializing as camel case
Shape keys where the first couple of characters were uppercase were
changed to become camel case. This obviously caused issues where shape
key blend values could not be loaded because the serialized key does not
correspond to a key in the TMorph's hash table.

The CamelCasePropertyNameContractResolver does not save dictionary keys
verbatim and treats them like json object properties and therefore
changes the keys to camel case. The worst part is that even if you
override the NamingStrategy to a CamelCaseNamingStrategy with
ProcessDictionaryKeys set to true, it STILL doesn't save dictionary keys
verbatim. You literally have to use a DefaultContractResolver and change
that NamingStrategy instead. Luckily, the damage seems minimal as most
shape keys are lowercase to begin with so it would not be noticeable by
the average user. TBody.SlotID enum values are all lowercase as well so
there's no damage there.