Fix gizmo click not detecting properly
The ray cast was not detecting clicks on drag handles for certain props
because those props had colliders. This caused the click check to return
a false positive where neither NGUI nor drag handles were clicked. This
solution checks for multiple colliders to see if any are NGUI or drag
handles. The real solution would be to sanitize props when they're
instantiated by removing their colliders but this is good for now.