Make pane headers more distinct
The pane headers were just toggles with larger text. Toggles are a
common control in the UI and despite the larger text, the header can
still be hard to distinguish. This change centers the text and adds a
dark background that spans with width of the panel. Arrows replace the
toggle radio icon to further separate the pane header from regular
toggle controls.