Make window background darker
A big problem with using GUI.skin.box style for windows is that the
background is too transparent and leads to low contrast for body text.
This change makes the windows use a dark purple texture. Slider track
needs a different texture as well because the slider track is already
very dark.

A more transparent version of the background is also added. This
background is used when the mouse is not hovering over the window for
those that would want to be able to see under the UI. This functionality
is inspired by ShapeKeyMaster.

Why purple? It's my favourite colour :3