Handle no hand items or attached accessories
Hand items and attached accessories are menu file items that are not
made available through public APIs in the game. Getting these items
would require querying all the arc files for them.
ShortMenuVanillaDatabase is a plugin that adds these items to the game's
menu file database because the plugin does not discriminate menu files
by MPN.