Add MeidoPhotoStudio source code analyzer
This is an analyzer custom made for MPS source code. Right now the only
thing that's analyzed are the translation keys.

Translation strings in MPS are retrieved through a Get method that takes
a category string and key string both of which are usually string
literals. This analyzer can verify if the arguments are contained in the
translation files.