Add GUIContent to controls
GUIContent is an object that represents the content of a control. Prior
to this change, GUIContent has been used sparingly and usually only
exposed the string and image members. This change now fully exposes
GUIContent for most of the controls and GUIContent or even GUIContent
derived classes can be used to store and represent the content of the
control.