Add LocalizableGUIContent
A big issue I've had with handling translations is that I needed
initialize a control with a translation and then later update the
control's text when the translation is reloaded. This means that there
can be copy paste issues or typos or controls that aren't retranslated.

This change is the core reason for exposing GUIContent in the controls.
What this does is updates the text member of the GUIContent when the
translation has been reloaded. This means that controls can just be
initialized once and the LocalizableGUIContent will handle updating the
text when the translation is reloaded.