Change method of changing bloom value
The way bloom was set would reactivate bloom despite the controller's
inactive state. This is most noticeable when loading a scene that has a
non-zero bloom value while it's disabled.

This change keeps track of the controller's current bloom value and sets
the game's bloom value when appropriate. The effect is now also set
active by default initializing with the game's current bloom setting.