Tweak header style
The header used a custom padding that placed it closer to other controls
below it as well as offsetting the text from the left. This change
restores the padding to the default for a label and uses ContentOffset
for the left side offset. The original intent for the padding is to have
the header sit on top of a line. The lines have been removed as the
style of the header evolved so this padding is not necessary anymore.