Improve dropdown label size calculation performance
The biggest bottleneck in opening a dropdown is calculating the width of
the dropdown window. Calculating the width requires iterating through
the whole dropdown list and finding the widest item when rendered. This
is a very expensive as it uses GUIStyle.CalcSize for every item. For
very large sets, 100k items for example, this causes the whole game to
freeze for a couple of seconds. Menu files usually surpass this number
and is where the freezing most likely occurs.

This change vastly improves performance by using GUIStyle.CalcSize to
calculate the sizes of the individual characters and caches them for use
in calculating the total width and height of a dropdown item manually.

I personally have 60k menu files and so the performance improvement is
not as apparent for me and my hardware but for other users that have
tested with 150k and a staggering 350k menu files, the difference is
huge.

This change does not scale very well when it comes to a test set of 1
million items and will freeze just the same although not to the same
extent.