Add pane group and sub-pane group
A pane group has a pane header and panes. A sub-pane group has a
sub-pane header and panes. A pane group encapsulate the behaviour of
drawing a header and the panes below it. A sub-pane group is drawn
similarly where there is a dark background encompassing the header and
the panes and black lines separating the panes.

A pane group has additional behaviour when a sub-pane group is added to
it. When a user clicks a pane group header while holding control it will
do the same thing as a header group closing the other headers. When the
pane group header is the only one open and is clicked while control is
held all the sub pane groups will be toggled. This behaviour is inspired
by blender.