Update NumericalTextField text regardless of focus
The initial idea was that having the text change while the text field
was being edited is pretty annoying. This would be fairly useful if
IMGUI handles focus better.

Handling of invalid input had to be tweaked a little to instead default
to the text field's current value instead as it would cause unnecessary
text field updates while the user is most definitely still focused and
typing. Slider needed to be tweaked as well to not update the text field
when a change comes from the text field.