Do not limit rotations while animation is playing
Rotation limits are applied even while the animation is playing while
not really doing anything visually. This means that despite the limited
joint appearing in a certain position, the joint is actually being moved
to a position within the limits of the rotation. This is apparent when a
user uses the MTAccelerator plugin where the limited joints appear to
flicker between the position the animation dictates and the position the
rotation limit dictates. A more severe visual bug that occurs, even
without MTAccelerator, is a prop that is attached to the limited joint
is visually at the point where the joint would be when limited while the
animation is playing. The prop does not event flicker in this case and
instead is just in an unexpected position and that is made apparent once
the animation is paused and the limited joint matches the position of
the prop.